Improving camera and loading tscn file
